services/ml-worker/src/detectors/image_detector.py
import io
import logging
from typing import Any, Dict, List, Optional

import numpy as np
from PIL import Image

from src.config import settings

logger = logging.getLogger(__name__)


class ImageDetector:
    """
    AI-generated image detection using ensemble of models.

    This is a mock implementation that demonstrates the detection pipeline.
    In production, this would load actual trained models.
    """

    def __init__(self):
        self.models_loaded = False
        self._load_models()

    def _load_models(self):
        """Load detection models."""
        logger.info("Loading detection models...")

        # In production, load actual models here:
        # self.deepfake_model = self._load_model("deepfake_v3.onnx")
        # self.gan_detector = self._load_model("gan_detector_v2.onnx")
        # self.diffusion_detector = self._load_model("diffusion_v1.onnx")

        self.models_loaded = True
        logger.info("Detection models loaded successfully")

    async def detect(self, image_data: bytes, options: Dict[str, Any]) -> Dict[str, Any]:
        """
        Run detection on an image.

        Args:
            image_data: Raw image bytes
            options: Detection options

        Returns:
            Detection result dictionary
        """
        # Load image
        image = Image.open(io.BytesIO(image_data))

        # Preprocess
        image = self._preprocess(image)

        # Run individual detectors
        detections = []

        # Primary classifier
        primary_result = await self._run_primary_classifier(image)
        detections.append(primary_result)

        # GAN detector
        gan_result = await self._run_gan_detector(image)
        detections.append(gan_result)

        # Diffusion detector
        diffusion_result = await self._run_diffusion_detector(image)
        detections.append(diffusion_result)

        # Frequency analysis
        freq_result = await self._run_frequency_analysis(image)
        detections.append(freq_result)

        # Calculate ensemble score
        ensemble_score = self._calculate_ensemble_score(detections)

        # Determine verdict
        verdict, risk_level = self._determine_verdict(ensemble_score)

        # Generate summary
        summary = self._generate_summary(verdict, detections)

        return {
            "verdict": verdict,
            "confidence": ensemble_score,
            "risk_level": risk_level,
            "summary": summary,
            "detections": detections,
            "ensemble_score": ensemble_score,
        }

    def _preprocess(self, image: Image.Image) -> Image.Image:
        """Preprocess image for detection."""
        # Convert to RGB if necessary
        if image.mode != "RGB":
            image = image.convert("RGB")

        # Resize if too large
        max_dim = settings.image_max_dimension
        if max(image.size) > max_dim:
            ratio = max_dim / max(image.size)
            new_size = (int(image.width * ratio), int(image.height * ratio))
            image = image.resize(new_size, Image.Resampling.LANCZOS)

        return image

    async def _run_primary_classifier(self, image: Image.Image) -> Dict[str, Any]:
        """Run primary CNN + ViT classifier."""
        # Mock implementation - in production, run actual model inference
        img_array = np.array(image)

        # Simulate detection based on image statistics
        mean_val = np.mean(img_array)
        std_val = np.std(img_array)

        # Mock confidence based on image properties
        confidence = self._mock_detection_score(mean_val, std_val)

        verdict = "ai_generated" if confidence > 0.5 else "authentic"

        return {
            "model": "primary_classifier_v3",
            "verdict": verdict,
            "confidence": confidence,
            "details": {
                "generation_type": "full_synthetic" if confidence > 0.7 else "unknown",
                "artifacts_detected": self._detect_artifacts(img_array),
            },
        }

    async def _run_gan_detector(self, image: Image.Image) -> Dict[str, Any]:
        """Run GAN fingerprint detector."""
        img_array = np.array(image)

        # Mock GAN detection
        confidence = self._mock_detection_score(np.mean(img_array), np.std(img_array), offset=0.1)

        generator = None
        if confidence > 0.6:
            # Mock generator identification
            generators = ["stylegan2", "stylegan3", "stable_diffusion", "midjourney", "dall_e_3"]
            generator = np.random.choice(generators)

        return {
            "model": "gan_detector_v2",
            "verdict": "ai_generated" if confidence > 0.5 else "authentic",
            "confidence": confidence,
            "details": {
                "likely_generator": generator,
                "fingerprint_match": confidence > 0.7,
            },
        }

    async def _run_diffusion_detector(self, image: Image.Image) -> Dict[str, Any]:
        """Run diffusion model detector."""
        img_array = np.array(image)

        confidence = self._mock_detection_score(np.mean(img_array), np.std(img_array), offset=-0.05)

        return {
            "model": "diffusion_detector_v1",
            "verdict": "ai_generated" if confidence > 0.5 else "authentic",
            "confidence": confidence,
            "details": {
                "model_family": "latent_diffusion" if confidence > 0.6 else None,
                "estimated_steps": "20-50" if confidence > 0.7 else None,
            },
        }

    async def _run_frequency_analysis(self, image: Image.Image) -> Dict[str, Any]:
        """Run frequency domain analysis (FFT/DCT)."""
        img_array = np.array(image.convert("L"))  # Grayscale

        # Compute FFT
        fft = np.fft.fft2(img_array)
        fft_shift = np.fft.fftshift(fft)
        magnitude = np.log(np.abs(fft_shift) + 1)

        # Analyze frequency distribution
        center = np.array(magnitude.shape) // 2
        high_freq_energy = np.mean(magnitude[center[0]-50:center[0]+50, center[1]-50:center[1]+50])

        # AI images often have different frequency characteristics
        confidence = min(1.0, max(0.0, (high_freq_energy - 5) / 10))

        return {
            "model": "frequency_analyzer_v1",
            "verdict": "ai_generated" if confidence > 0.5 else "authentic",
            "confidence": confidence,
            "details": {
                "fft_score": float(confidence),
                "spectral_anomaly": confidence > 0.7,
            },
        }

    def _calculate_ensemble_score(self, detections: List[Dict]) -> float:
        """Calculate weighted ensemble score from all detectors."""
        weights = {
            "primary_classifier_v3": 0.35,
            "gan_detector_v2": 0.25,
            "diffusion_detector_v1": 0.25,
            "frequency_analyzer_v1": 0.15,
        }

        total_weight = 0
        weighted_sum = 0

        for detection in detections:
            model = detection["model"]
            weight = weights.get(model, 0.1)
            weighted_sum += detection["confidence"] * weight
            total_weight += weight

        return weighted_sum / total_weight if total_weight > 0 else 0.5

    def _determine_verdict(self, confidence: float) -> tuple[str, str]:
        """Determine verdict and risk level from confidence score."""
        if confidence > 0.85:
            return "ai_generated", "high"
        elif confidence > 0.65:
            return "likely_ai", "medium"
        elif confidence > 0.35:
            return "inconclusive", "medium"
        elif confidence > 0.15:
            return "likely_authentic", "low"
        else:
            return "authentic", "low"

    def _generate_summary(self, verdict: str, detections: List[Dict]) -> str:
        """Generate human-readable summary of detection results."""
        summaries = {
            "ai_generated": "This image shows strong indicators of AI generation. Multiple detection models agree on synthetic origin.",
            "likely_ai": "This image shows probable signs of AI generation. Some indicators suggest synthetic content.",
            "inconclusive": "Detection results are inconclusive. The image shows mixed signals that don't clearly indicate authentic or synthetic origin.",
            "likely_authentic": "This image appears likely to be authentic. Few AI-generated indicators detected.",
            "authentic": "This image shows strong indicators of being an authentic photograph with no signs of AI generation.",
        }

        # Add specific findings
        ai_models = [d for d in detections if d["verdict"] == "ai_generated"]
        if ai_models:
            generators = [
                d["details"].get("likely_generator")
                for d in ai_models
                if d["details"].get("likely_generator")
            ]
            if generators:
                return f"{summaries[verdict]} Likely source: {generators[0]}."

        return summaries.get(verdict, "Unable to determine image authenticity.")

    def _mock_detection_score(self, mean: float, std: float, offset: float = 0) -> float:
        """Generate mock detection score for testing."""
        # This creates semi-random but consistent scores based on image properties
        base_score = (np.sin(mean / 50) + np.cos(std / 30)) / 2 + 0.5
        return min(1.0, max(0.0, base_score + offset))

    def _detect_artifacts(self, img_array: np.ndarray) -> List[str]:
        """Detect visual artifacts that might indicate AI generation."""
        artifacts = []

        # Check for texture consistency issues
        local_stds = []
        for i in range(0, img_array.shape[0] - 32, 32):
            for j in range(0, img_array.shape[1] - 32, 32):
                patch = img_array[i:i+32, j:j+32]
                local_stds.append(np.std(patch))

        if len(local_stds) > 0 and np.std(local_stds) < 15:
            artifacts.append("texture_uniformity")

        # Check for color anomalies
        if img_array.ndim == 3:
            color_ranges = [np.ptp(img_array[:, :, c]) for c in range(3)]
            if min(color_ranges) < 100:
                artifacts.append("limited_color_range")

        return artifacts

    async def generate_heatmap(
        self, image_data: bytes, result: Dict[str, Any]
    ) -> Optional[bytes]:
        """Generate a heatmap showing suspicious regions."""
        try:
            image = Image.open(io.BytesIO(image_data))
            img_array = np.array(image.convert("RGB"))

            # Create simple gradient heatmap for demonstration
            # In production, this would highlight actual suspicious regions
            h, w = img_array.shape[:2]
            heatmap = np.zeros((h, w), dtype=np.float32)

            # Create radial gradient as mock heatmap
            y, x = np.ogrid[:h, :w]
            center_y, center_x = h // 2, w // 2
            heatmap = 1 - np.sqrt((x - center_x) ** 2 + (y - center_y) ** 2) / (
                np.sqrt(center_x ** 2 + center_y ** 2)
            )
            heatmap = np.clip(heatmap * result["confidence"], 0, 1)

            # Apply colormap (red = suspicious)
            heatmap_colored = np.zeros((h, w, 4), dtype=np.uint8)
            heatmap_colored[:, :, 0] = (heatmap * 255).astype(np.uint8)  # Red
            heatmap_colored[:, :, 3] = (heatmap * 180).astype(np.uint8)  # Alpha

            # Overlay on original image
            overlay = Image.fromarray(heatmap_colored, mode="RGBA")
            original_rgba = image.convert("RGBA")
            composite = Image.alpha_composite(original_rgba, overlay)

            # Save to bytes
            buffer = io.BytesIO()
            composite.save(buffer, format="PNG")
            return buffer.getvalue()

        except Exception as e:
            logger.error(f"Failed to generate heatmap: {e}")
            return None
