services/ml-worker/pyproject.toml
[project]
name = "forensivision-ml-worker"
version = "1.0.0"
description = "ForensiVision ML Inference Worker"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "onnxruntime-gpu>=1.17.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "aio-pika>=9.3.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "scikit-image>=0.22.0",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
